Compile error resolved

+++ src/main/java/com/iemr/admin/repository/user/UserLoginRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.admin.data.user.M_User;
 
 @Repository
-public interface UserLoginRepo extends CrudRepository<M_User, Long> {
+public interface UserLoginRepo extends CrudRepository<M_User, Integer> {
 
 	@Query(" SELECT u FROM M_User u WHERE u.userID = :userID AND u.Deleted = false ")
 	public M_User getUserByUserID(@Param("userID") Long userID);
